snowflake_data_profiler/requirements.txt
pandas==1.0.5
pandas-profiling==2.8.0
Flask==1.1.2
Flask-WTF==0.14.3
Jinja2==2.11.2
snowflake-connector-python==2.2.8
WTForms==2.3.1


